feature(dws): 适配dws数据源
Closes #148156

+++ iengine/iengine-app/src/main/java/io/tapdata/flow/engine/V2/node/hazelcast/data/pdk/HazelcastTargetPdkDataNode.java
@@ -150,6 +150,9 @@ public class HazelcastTargetPdkDataNode extends HazelcastTargetPdkBaseNode {
 						if (null != funcAspect) funcAspect.setThrowable(e);
 						throw e;
 					}
+					if (StringUtils.isNotBlank(tableId) && StringUtils.equalsAny(tableId, ExactlyOnceUtil.EXACTLY_ONCE_CACHE_TABLE_NAME)) {
+						continue;
+					}
 					dropTable(existsDataProcessEnum, tableId);
 					boolean createdTable = createTable(tapTable);
 					clearData(existsDataProcessEnum, tableId);
